MNT: move test script over to unchecked interactive_tests
